+++ src/i18n/types.d.ts
@@ -7,6 +7,7 @@ import type en_warnings from "./locales/en/warnings.json";
 import type en_navigation from "./locales/en/navigation.json";
 import type en_scopes from "./locales/en/scopes.json";
 import type en_filter from "./locales/en/filter.json";
+import type en_statusIcons from "./locales/en/statusIcons.json";
 
 declare module "i18next" {
   interface CustomTypeOptions {
@@ -21,6 +22,7 @@ declare module "i18next" {
       navigation: typeof en_navigation;
       scopes: typeof en_scopes;
       filter: typeof en_filter;
+      statusIcons: typeof en_statusIcons;
     };
   }
 }